Move server views to `lib/server/views`
Rename the index view to layout. Makes more sense.
