+++ lib/server/routes.js
@@ -1,5 +1,5 @@
 import express from 'express'
-import views from '../views'
+import views from './views'
 import api from '../api'
 
 const router = express.Router()
@@ -10,7 +10,7 @@ router.get('/wiki/:title', (req, res) => {
   var escapedTitle = title.replace(/_/g, ' ')
   api.lead(title)
     .then((content) =>
-      res.end(views.index({
+      res.end(views.layout({
         title: escapedTitle,
         body: views.articleSummary({
           title: escapedTitle,
@@ -25,7 +25,7 @@ router.get('/wiki/full/:title', (req, res) => {
   var escapedTitle = title.replace(/_/g, ' ')
   api.full(title)
     .then((content) =>
-      res.end(views.index({
+      res.end(views.layout({
         title: escapedTitle,
         body: views.article({
           title: escapedTitle,
