+++ lib/server/views/index.js
@@ -2,13 +2,13 @@ var fs = require('fs')
 var ejs = require('ejs')
 
 module.exports = {
-  index: template('index'),
+  layout: template('layout'),
   articleSummary: template('article-summary'),
   article: template('article')
 }
 
 function template (name) {
-  var str = fs.readFileSync(`${__dirname}/views/${name}.ejs`).toString()
+  var str = fs.readFileSync(`${__dirname}/${name}.ejs`).toString()
   return ejs.compile(str)
 }
